+++ tests/testthat/test-theil.R
@@ -1,0 +1,87 @@
+context('theil')
+
+# -------------- #
+# theil testthat #
+# -------------- #
+
+test_that('theil throws error with invalid arguments', {
+  # Unavailable geography
+  expect_error(
+    theil(
+      geo_small = 'zcta',
+      state = 'DC',
+      year = 2020,
+      subgroup = 'NHoLB',
+      quiet = TRUE
+    )
+  )
+  expect_error(
+    theil(
+      geo_large = 'block group',
+      state = 'DC',
+      year = 2020,
+      subgroup = 'NHoLB',
+      quiet = TRUE
+    )
+  )
+  
+  # Unavailable year
+  expect_error(
+    theil(
+      state = 'DC',
+      year = 2005,
+      subgroup = 'NHoLB',
+      quiet = TRUE
+    )
+  )
+  
+  # Unavailable subgroup
+  expect_error(
+    theil(
+      state = 'DC',
+      year = 2020,
+      subgroup = 'terran',
+      quiet = TRUE
+    )
+  )
+  
+  skip_if(Sys.getenv('CENSUS_API_KEY') == '')
+  
+  # Incorrect state
+  expect_error(
+    theil(
+      state = 'AB',
+      year = 2020,
+      subgroup = 'NHoLB',
+      quiet = TRUE
+    )
+  )
+  
+})
+
+test_that('theil works', {
+  skip_if(Sys.getenv('CENSUS_API_KEY') == '')
+  
+  expect_silent(theil(
+    state = 'DC',
+    year = 2020,
+    subgroup = c('NHoLB', 'HoLB'),
+  ))
+  
+  expect_silent(
+    theil(
+      state = 'DC',
+      year = 2020,
+      subgroup = 'NHoLB',
+      quiet = TRUE
+    )
+  )
+  
+  expect_silent(theil(
+    state = 'DC',
+    year = 2020,
+    subgroup = c('NHoLB', 'HoLB'),
+    quiet = TRUE
+  ))
+  
+})